Rename `projection` package to `stream`.
