+++ core-java/src/test/java/org/spine3/server/BoundedContextShould.java
@@ -36,10 +36,10 @@ import org.spine3.server.aggregate.AggregateShould;
 import org.spine3.server.error.UnsupportedCommandException;
 import org.spine3.server.procman.ProcessManager;
 import org.spine3.server.procman.ProcessManagerRepository;
-import org.spine3.server.projection.StreamProjection;
-import org.spine3.server.projection.StreamProjectionRepository;
 import org.spine3.server.storage.StorageFactory;
 import org.spine3.server.storage.memory.InMemoryStorageFactory;
+import org.spine3.server.stream.StreamProjection;
+import org.spine3.server.stream.StreamProjectionRepository;
 import org.spine3.test.project.ProjectId;
 import org.spine3.test.project.event.ProjectCreated;
 import org.spine3.test.project.event.ProjectStarted;
